Moving prune script to appropriate location
	typechange: prune_modulestore.py
	copied:     prune_modulestore.py -> ../tubular/scripts/prune_modulestore.py
